Revert "Use FontAwesome's CDN instead of hosting it"
This reverts commit ec299ae478a3fa7106c7f7189579f394d1ce9fc5.
FontAwesome's CDN does not support SRI, which is a dealbreaker as we are trying to tighten up security in preparation of online payments
Furthermore, the async attirbute can be used to improve performance past their CDN, since that was their only advantage (+ caching). We also don't have to open a new connection which shaves off latency.